requirements.txt
# RAG System Dependencies - production

# Vector Database
faiss-cpu>=1.7.4,<2.0
chromadb>=0.4.0,<1.0

# Machine Learning Models
torch>=2.0.0,<3.0
torchvision>=0.15.0,<1.0
transformers>=4.30.0,<5.0
sentence-transformers>=2.2.0,<4.0
clip-by-openai>=1.0,<2.0

# Image Processing
opencv-python>=4.8.0,<5.0
Pillow>=10.0.0,<12.0
numpy>=1.24.0,<3.0

# API Framework
fastapi>=0.100.0,<1.0
uvicorn>=0.23.0,<1.0
pydantic>=2.0.0,<3.0

# Web Interface
flask>=2.3.0,<4.0
jinja2>=3.1.0,<4.0

# Database
sqlalchemy>=2.0.0,<3.0

# Authentication & Security
python-jose[cryptography]>=3.3.0,<4.0
passlib[bcrypt]>=1.7.4,<2.0
bcrypt>=4.0.0,<5.0
python-multipart>=0.0.6,<1.0

# System Monitoring
psutil>=5.9.0,<7.0

# Caching
redis>=4.6.0,<6.0

# Utilities
requests>=2.31.0,<3.0
python-dotenv>=1.0.0,<2.0
tqdm>=4.65.0,<5.0

# Analytics
matplotlib>=3.7.0,<4.0
seaborn>=0.12.0,<1.0
plotly>=5.15.0,<6.0

# Export
pandas>=2.0.0,<3.0
openpyxl>=3.1.0,<4.0

# Rate Limiting
slowapi>=0.1.9,<1.0
